softoff now in degrees (double) changes in verge not being sent to spike2 (e.g. fx,fy), now fixed.

+++ binoclean/stimuli.h
@@ -412,7 +412,7 @@ struct EXPERIMENT{
     int stimmode,ncalc,noverflow; 
     int fastextras,driveadapter;
     float fixpos[2];
-    int softoff[6];
+    double softoff[6];
     int hemisphere,electrode;
     double minval, fastincr;
     int nfast;
